add support for BinPackFirstFitWeight
With BinPackFirstFitWeight > 0, it will make diego-cells with a lower
BOSH instance index number more attractive to deploy LRPs to.
Diego-cells will be filled up more instead spreading the containers
accross all diego-cells equally.

* Why?

Currently we have 64GB memory diego-cells. We need to have the
possibility for our customers to deploy 16G containers at all times.
This means that we need to have at least 1 diego-cell with 16G memory
available at all times.

Previously on average 25% (16/64) of the memory on all diego-cells was
not used, because every time an LRP is scheduled for deployment, it
picks the diego-cell with most memory/disk/pids available.

With this change the BOSH instance index number adds some weight to the
scheduling algorithm, which will push containers to diego-cells with a
lower instance index number when the resources (memory, disk, pids) are
available.